+++ packages/data-model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/data-model
 
+## 1.3.15
+
+### Patch Changes
+
+- feat: refine snap related api
+
 ## 1.3.14
 
 ### Patch Changes